lact_nvs/requirements.txt
torch>=2.4.0
torchvision
einops
lpips
omegaconf
Pillow
imageio
numpy
transformers
huggingface-hub